Delegate library references handling from pull command to pull manager
